Replace the Ruby tests with mocha
This PR proposals a potential path for replacing the Ruby based
test runner with the mocha which is a node testing pseudo standard.

This project has a hard dependency node-sass. Given users of this
project have chosen node-sass a dependency on Ruby is a barrier.
Given that the Ruby dependency is for testing, this reduces the
likely hood that users will write tests for their modules.

It's worth mentioning that Ruby does not come preinstalled on Windows
and many *nix distributions like it does with OSX.